Updated _parse_function
